+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### 1.5.1 (Next)
 
-* ...
+* [#65](https://github.com/ruby-grape/grape-active_model_serializers/pull/65): Added Danger, PR linter - [@dblock](https://github.com/dblock).
+* Your contribution here.
 
 ### 1.5.0 (August 24, 2016)